goodMorning

+++ src/app/item-list/item-list.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormBuilder } from '@angular/forms';
 
 import { StoreService } from '../../store.service';
+import { ListItemComponent } from '../list-item/list-item.component'
 
 @Component({
   selector: 'app-item-list',
@@ -26,17 +27,26 @@ export class ItemListComponent implements OnInit {
   }
 
   onSubmit(data) {
-    const item = {
-      text: data.inputText,
-      created: new Date(),
-      complete: false,
-      id: new Date().getTime()
+    if(data.inputText)
+      {
+      const item = {
+        text: data.inputText,
+        created: new Date(),
+        complete: false,
+        id: new Date().getTime()
+      }
+      this.storeService.addItem(item);
+      this.addItemInput.reset();
     }
-    this.storeService.addItem(item);
-    this.addItemInput.reset();
   }
 
   onRemove(id) {
-    this.storeService.removeItem(id)
+    var r = confirm("remove item from list?");
+    if (r == true) {
+      this.storeService.removeItem(id)
+    } else {
+      return
+    }
   }
+
 }
